Refactoring to reduce duplication between Deposit/Withdrawal. Not comfortable with the inheritance of Transaction, but it is clean
