--- MakersAcademy.Bank/ITransaction.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace MakersAcademy.Bank
-{
-    public interface ITransaction
-    {
-        DateTime DateTime { get; }
-        int DepositAmount { get; }
-        int WithdrawalAmount { get; }
-        int AdjustBalance(int previousBalance);
-    }
-}
