+++ MakersAcademy.Bank/Withdrawal.cs
@@ -2,22 +2,10 @@ using System;
 
 namespace MakersAcademy.Bank
 {
-    public class Withdrawal : ITransaction
+    public class Withdrawal : Transaction
     {
-        public DateTime DateTime { get; }
-        public int DepositAmount { get; }
-        public int WithdrawalAmount { get; }
-
-        public int AdjustBalance(int previousBalance)
+        public Withdrawal(DateTime dateTime, int amount) : base(dateTime, withdrawalAmount: amount, depositAmount: 0)
         {
-            return previousBalance + DepositAmount - WithdrawalAmount;
-        }
-
-        public Withdrawal(DateTime dateTime, int amount)
-        {
-            DateTime = dateTime;
-            WithdrawalAmount = amount;
-            DepositAmount = 0;
         }
     }
 }
